# docs/api-contract.yaml
openapi: 3.0.3
info:
  title: Restart API
  version: 1.0.0
  description: |
    ## Overview
    Restart is a multi-tenant SaaS platform for managing instructor-led training programmes.
    Each instructor (tenant) operates in an isolated PostgreSQL schema — their data is
    completely separate from every other tenant.

    ## Two Scopes
    The API has two distinct actor scopes:

    - **ADMIN** — Internal staff. Data lives in the `public` schema. Manages tenant creation,
      instructor provisioning, tracks, weekly targets, and has read access across all tenants.
    - **INSTRUCTOR** — One per tenant. Data lives in the tenant's own schema. Manages their
      own cohorts, learners, and attendance records.

    ## Authentication
    All protected endpoints require a `Bearer` JWT in the `Authorization` header.

    - Admin tokens are obtained from `POST /api/auth/admin/login`.
    - Instructor tokens are obtained from `POST /api/auth/login` with an `X-Tenant-ID` header
      identifying which tenant schema to authenticate against.

    ### Why X-Tenant-ID on login?
    PostgreSQL schema-per-tenant means "john" in `tenant_a` and "john" in `tenant_b` are
    completely different users. The `X-Tenant-ID` header tells the backend which schema to
    search in during authentication. After login, the tenant is embedded in the JWT itself,
    so subsequent requests do **not** need `X-Tenant-ID` — the token carries it.

    ## Standard Response Envelope
...
    ```json
    {
...
    }
...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /api/tenants/{schemaName}:
    get:
      tags: [Tenants]
      summary: Get a single tenant
      parameters:
        - $ref: '#/components/parameters/SchemaNamePath'
      responses:
        '200':
          description: Tenant found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseTenantResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── ADMIN › INSTRUCTORS ───────────────────────────────────────────────────
  /api/admin/instructors:
    post:
      tags: [Admin › Instructors]
      summary: Provision an instructor account
      description: |
        Creates a user record in the tenant's `users` table (not the public schema).
        The instructor can then log in via `POST /api/auth/login` with `X-Tenant-ID`.
        Multiple instructor accounts can exist per tenant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstructorProvisionRequest'
            example:
              schemaName: kwame_mensah
              username: kwame
              password: securePass123
      responses:
        '201':
          description: Instructor provisioned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseInstructorResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/instructors/{schemaName}:
    get:
      tags: [Admin › Instructors]
      summary: List instructor accounts in a tenant
      parameters:
        - $ref: '#/components/parameters/SchemaNamePath'
      responses:
        '200':
          description: List of instructors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseInstructorList'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/instructors/{schemaName}/{username}:
    delete:
      tags: [Admin › Instructors]
      summary: Remove an instructor account
      parameters:
        - $ref: '#/components/parameters/SchemaNamePath'
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── ADMIN › LEARNERS ──────────────────────────────────────────────────────
...
  /api/admin/tenants/{schemaName}/attendance:
    get:
      tags: [Admin › Attendance]
      summary: List all attendance records for a tenant
      parameters:
        - $ref: '#/components/parameters/SchemaNamePath'
      responses:
        '200':
          description: Attendance list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAttendanceList'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/tenants/{schemaName}/attendance/cohort/{cohortId}:
    get:
      tags: [Admin › Attendance]
      summary: Filter attendance by cohort
      parameters:
        - $ref: '#/components/parameters/SchemaNamePath'
        - name: cohortId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Filtered attendance list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAttendanceList'
  /api/admin/tenants/{schemaName}/attendance/{id}:
    get:
      tags: [Admin › Attendance]
      summary: Get a single attendance record
      parameters:
        - $ref: '#/components/parameters/SchemaNamePath'
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Attendance record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAttendanceResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── ADMIN › WEEKLY TARGETS ───────────────────────────────────────────────
  /api/admin/weekly-targets:
    post:
      tags: [Admin › Weekly Targets]
      summary: Create a weekly target
      description: |
        Associates a week number (WEEK_1 through WEEK_10) with a list of lab tasks and
        knowledge checks for a given track. The combination of `(trackId, weekNumber)`
        must be unique. Instructor cohorts reference a `trackId` to pull the right schedule.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeeklyTargetRequest'
      responses:
        '201':
          description: Weekly target created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseWeeklyTargetResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/admin/weekly-targets/{id}:
    get:
      tags: [Admin › Weekly Targets]
      summary: Get a weekly target by ID
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Weekly target
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseWeeklyTargetResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Admin › Weekly Targets]
      summary: Update a weekly target
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeeklyTargetRequest'
      responses:
        '200':
          description: Updated weekly target
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseWeeklyTargetResponse'
    delete:
      tags: [Admin › Weekly Targets]
      summary: Delete a weekly target
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'
  /api/admin/weekly-targets/track/{trackId}:
    get:
      tags: [Admin › Weekly Targets]
      summary: Get all weekly targets for a track
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Weekly targets for the track
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseWeeklyTargetList'
  /api/admin/weekly-targets/track/{trackId}/week/{weekNumber}:
    get:
      tags: [Admin › Weekly Targets]
      summary: Get a specific week's targets
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: weekNumber
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/WeekNumber'
      responses:
        '200':
          description: Weekly target for the given track and week
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseWeeklyTargetResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── ADMIN › AUDIT LOGS ───────────────────────────────────────────────────
  /api/admin/audit-logs:
    get:
      tags: [Admin › Audit Logs]
      summary: Query the audit trail
      description: |
        Returns a paginated list of audit log entries. Apply at most **one** filter at a time
        (precedence: `actorUsername` → `tenantId` → `resourceType` → `action`). Results are
        always sorted newest-first.

        Every endpoint call is logged regardless of whether it has a specific `@Auditable`
        action — the HTTP method, path, and status are always captured. Endpoints annotated
        with `@Auditable` additionally carry `action`, `resourceType`, `resourceId`, and
        `tenantId`.
      parameters:
        - name: actorUsername
          in: query
          schema:
            type: string
          description: Filter by the username of the actor who made the request
        - name: tenantId
          in: query
          schema:
            type: string
          description: Filter by tenant schema name (e.g. `kwame_mensah`)
        - name: resourceType
          in: query
          schema:
            type: string
            enum: [LEARNER, COHORT, ATTENDANCE, INSTRUCTOR, WEEKLY_TARGET, TENANT]
        - name: action
          in: query
          schema:
            type: string
          description: Exact action string (e.g. `CREATE_LEARNER`, `DELETE_COHORT`)
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated audit log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAuditLogPage'

  # ── INSTRUCTOR › COHORTS ─────────────────────────────────────────────────
...
  /api/instructor/attendance/cohort/{cohortId}:
    get:
      tags: [Instructor › Attendance]
      summary: Filter attendance by cohort
      parameters:
        - name: cohortId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Filtered attendance list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAttendanceList'
  /api/instructor/attendance/{id}:
    get:
      tags: [Instructor › Attendance]
      summary: Get a single attendance record
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Attendance record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAttendanceResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Instructor › Attendance]
      summary: Update an attendance record
      description: Re-validates all participant learner IDs on every update.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceRequest'
      responses:
        '200':
          description: Attendance updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAttendanceResponse'
    delete:
      tags: [Instructor › Attendance]
      summary: Delete an attendance record
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'

# ─────────────────────────────────────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by `/api/auth/admin/login` or `/api/auth/login`.
        Claims: `sub` (username), `role` (ADMIN | INSTRUCTOR), `tenantId` (schema name, null for admin).
        Expiry: 24 hours.
  parameters:
    SchemaNamePath:
      name: schemaName
      in: path
      required: true
      description: |
        The PostgreSQL schema name for the tenant (e.g. `kwame_mensah`).
        Derived from the instructor's name via `SchemaUtil.toSchemaName()`:
        lowercase, non-alphanumeric chars → underscore, collapse multiple underscores.
      schema:
        type: string
        example: kwame_mensah
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
...
    InstructorProvisionRequest:
      type: object
      required: [schemaName, username, password]
      properties:
        schemaName:
          type: string
          example: kwame_mensah
        username:
          type: string
          example: kwame
        password:
          type: string
          format: password
          minLength: 8
          example: securePass123
...
    AttendanceRequest:
      type: object
      required: [cohortId, sessionDate, duration, participants]
      properties:
        cohortId:
          type: integer
          format: int64
          example: 1
        sessionDate:
          type: string
          format: date
          example: "2026-05-02"
        duration:
          type: integer
          description: Total session duration in minutes.
          minimum: 1
          example: 120
        participants:
          type: array
          minItems: 1
          description: |
            Each entry records one learner's presence. The `learnerId` must exist
            in the tenant's `learners` table — the API validates all IDs before saving.
            Per-participant `duration` captures how long that learner was actually present.
          items:
            $ref: '#/components/schemas/ParticipantEntry'
    ParticipantEntry:
      type: object
      required: [learnerId, duration]
      properties:
        learnerId:
          type: integer
          format: int64
          example: 42
        duration:
          type: integer
          description: Minutes the learner was present in this session.
          minimum: 1
          example: 90
    WeeklyTargetRequest:
      type: object
      required: [trackId, weekNumber, labs, knowledgeChecks]
      properties:
        trackId:
          type: integer
          format: int64
          example: 1
        weekNumber:
          $ref: '#/components/schemas/WeekNumber'
        labs:
          type: array
          items:
            type: string
          example: ["Lab 1: Introduction to Variables", "Lab 2: Control Flow"]
        knowledgeChecks:
          type: array
          items:
            type: string
          example: ["What is a variable?", "Explain the difference between if and switch"]
    WeekNumber:
      type: string
      enum:
        - WEEK_1
        - WEEK_2
        - WEEK_3
        - WEEK_4
        - WEEK_5
        - WEEK_6
        - WEEK_7
        - WEEK_8
        - WEEK_9
        - WEEK_10
      description: |
        Programme week. The combination `(trackId, weekNumber)` is UNIQUE — only one
        weekly target record can exist per track per week.

  # ── RESPONSE BODIES ──────────────────────────────────────────────────────
...
    TenantResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        schemaName:
          type: string
          description: The PostgreSQL schema name — use this as `X-Tenant-ID` for instructor login.
          example: kwame_mensah
        instructorName: { type: string, example: Kwame Mensah }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
    InstructorResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        schemaName: { type: string }
        role: { type: string, example: INSTRUCTOR }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    AttendanceResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        cohortId: { type: integer, format: int64 }
        sessionDate: { type: string, format: date }
        duration: { type: integer, description: Session duration in minutes }
        participants:
          type: array
          items:
            $ref: '#/components/schemas/ParticipantEntry'
        createdAt: { type: string, format: date-time }
    WeeklyTargetResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        trackId: { type: integer, format: int64 }
        weekNumber:
          $ref: '#/components/schemas/WeekNumber'
        labs:
          type: array
          items: { type: string }
        knowledgeChecks:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
    AuditLogResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        actorUsername: { type: string }
        actorRole: { type: string, enum: [ADMIN, INSTRUCTOR] }
        tenantId: { type: string, nullable: true, description: Schema name of the affected tenant }
        action: { type: string, nullable: true, description: "e.g. CREATE_LEARNER, DELETE_COHORT" }
        resourceType: { type: string, nullable: true }
        resourceId: { type: integer, format: int64, nullable: true }
        details: { type: string, nullable: true }
        ipAddress: { type: string }
        httpMethod: { type: string, example: POST }
        endpointPath: { type: string, example: /api/instructor/learners }
        httpStatus: { type: integer, example: 201 }
        createdAt: { type: string, format: date-time }
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        data: { nullable: true, example: null }

    # ── ENVELOPE WRAPPERS ─────────────────────────────────────────────────
...
    ApiResponseTenantResponse:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TenantResponse'
...
    ApiResponseInstructorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/InstructorResponse'
    ApiResponseInstructorList:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/InstructorResponse'
...
    ApiResponseAttendanceResponse:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AttendanceResponse'
    ApiResponseAttendanceList:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/AttendanceResponse'
    ApiResponseWeeklyTargetResponse:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/WeeklyTargetResponse'
    ApiResponseWeeklyTargetList:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/WeeklyTargetResponse'
    ApiResponseAuditLogPage:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: object
              description: Spring Data Page wrapper
              properties:
                content:
                  type: array
                  items:
                    $ref: '#/components/schemas/AuditLogResponse'
                totalElements: { type: integer }
                totalPages: { type: integer }
                size: { type: integer }
                number: { type: integer, description: Current page (0-indexed) }
    ApiEnvelope:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        data: {}
  responses:
...
    BadRequest:
      description: Validation error or bad input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "phone: Phone must be a valid 10-digit Ghanaian number"
            data: null
    NotFound:
      description: Resource or tenant not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Tenant not found: unknown_schema"
            data: null
    EmptySuccess:
      description: Action completed, no data returned
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiEnvelope'
          example:
            success: true
            message: Success
            data: null